aidl: light: manages backlights with max brighness level greater than 255
Some devices (e.g.: OnePlus 5T) have a display with a maximum brightness
greater than 255: it can be 1024, 2048 or even 4096.

In this case, we need to scale the brighness in a linear way as Android
only supports a 0-255 scale.

This is the same mechanism used on the previous not-AIDL HAL.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>